+++ src/Pixes/pix_posterize.cpp
@@ -37,14 +37,14 @@ pix_posterize :: ~pix_posterize()
 
 void pix_posterize :: factorMess(float f)
 {
-    factor = (int)(f*255.f);
-    setPixModified();
+  factor = static_cast<int>(f*255.f);
+  setPixModified();
 }
 
 void pix_posterize :: limitMess(float l)
 {
-    limit = (int)l;
-    setPixModified();
+  limit = static_cast<int>(l);
+  setPixModified();
 }
 
 /////////////////////////////////////////////////////////
@@ -139,10 +139,10 @@ class_addmethod(classPtr, reinterpret_cast<t_method>(&pix_posterize::limitMessCa
 
 void pix_posterize :: factorMessCallback(void *data, t_floatarg size)
 {
-    GetMyClass(data)->factorMess((float)size);
+    GetMyClass(data)->factorMess(size);
 }
 
 void pix_posterize :: limitMessCallback(void *data, t_floatarg size)
 {
-    GetMyClass(data)->limitMess((float)size);
+    GetMyClass(data)->limitMess(size);
 }
